ENH: Convert itkCompensatedSummationTest2 to itkCompensatedSummationGTest
